Move weather types to /types
